requirements.txt
camelsplit==1.0.1
certifi==2021.10.8
datasketch==1.5.7
elastic-transport==8.1.1
elasticsearch==7.8.1
future==0.18.2
graphviz==0.19.1
joblib==1.1.0
numpy==1.22.3
pandas==1.4.1
py4j==0.10.9.3
pyaml==21.10.1
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
threadpoolctl==3.1.0
urllib3==1.26.9

tqdm~=4.63.1
zlib~=1.2.11
feast~=0.19.3